Finish
All is set, the only thing are to fix some bugs. like having the dot spawns withing the lock-body circle, having the paddle move clockwise and anticlockwise depending on where the dot is, passible take away the animation and only keep the color change for a miss-dot or a level clears. more bugs are to be found and adjusted.